requirements.txt
usb-monitor
pycryptodome
psutil
